refactor: organize codebase structure and clean up technical debt
- Remove build artifacts from version control (29MB reduction)
- Organize configuration files into config/environments/ structure
- Restructure core scanning functionality into internal/scanning package
- Update all import references for new package structure
- Add comprehensive package documentation for scanning module
- Create structured TODO tracking for remaining technical debt
- Update documentation and scripts for new file locations
- Maintain full backward compatibility

This refactoring improves code organization, reduces repository size,
and provides better separation of concerns without breaking changes.
