+++ config/environments/config.minimal.yaml
