+++ internal/scanning/types.go
@@ -1,4 +1,4 @@
-package internal
+package scanning
 
 import (
 	"fmt"
