+++ internal/scanning/xml.go
@@ -1,4 +1,4 @@
-package internal
+package scanning
 
 import (
 	"encoding/xml"
