+++ internal/scanning/xml_test.go
@@ -1,4 +1,4 @@
-package internal
+package scanning
 
 import (
 	"os"
